Update api-register-agent.ps1
I've added a couple of useful lines to this script after finding a couple of issues we encountered when trying to migrate existing agents to a new IP address.

The suggested changes I have made:

Script will add the manager IP address to the config file by appending it to the last line.

Script will test it can reach the manager, I found that if the initial IE first run dialogue had not been run through it would stop the GET / POST actions.

Script will then check if the agent exists in the manager and if so retrieve the agent key if it already exists in the system and configure the agent appropriately. 
 
Hope this is useful!